Refactor: move view_file implementations to their own subdirectory.
