+++ src/view_file/Makefile.am
@@ -1,0 +1,5 @@
+module_view_file = \
+	%D%/view_file_icon.c	\
+	%D%/view_file_icon.h	\
+	%D%/view_file_list.c	\
+	%D%/view_file_list.h
