+++ src/view_file/view_file_icon.c
